fix: dBFS calculation is similar to pydub
 move chunks creation and dBFS to the proper three
